Add BlockExternalized and PreImageReceived handler events to Agora
Add event handlers to be used when the block is externalized and when receiving the preimage.
This is an implementation of the event handler and has not yet been starting with block reception.
Also, with the preimage reception,

Convert data into JSON format is sent to POST using RestInterfaceClient.

TestNetworkManager does not support it.