Bug fix: Make the validators of Genesis reveal pre-images timely
I fixed using the wrong type (ulong instead of uint) when creating
a pre-image hash. The bug caused the failure of revelation of pre-images of
the validators of the Genesis block. And I added a network test for restoring
the enrollment information of a node and revealing pre-images of `Enrollment`s
of the Genesis block.