Bug fix: Avoid calling receivePreimage endlessly
If a FullNode accept an pre-image although the distance is equal to that of
the previous one, it sends the pre-image to the network endlessly. You can
see the reason why it does in the following code of the `receivePreimage`
function of the `FullNode.d`.
```
if (this.enroll_man.addPreimage(preimage))
    this.network.sendPreimage(preimage);
```
So, I make FullNodes not accept the pre-image by returning the string of the
reason why the pre-image is invalid.