Remove GenesisTransaction function
Because we can have different kind of genesis blocks, we cannot offer an unified 'GenesisTransaction' API.
This function originally meant 'the only transaction in the genesis block',
and it's meaning recently changed to 'the spendable output of genesis'.
But for this, using a more programmatic approach, such as what 'genesisSpendable'
is doing makes much more sense and is more robust.
Since make of the places which use GenesisTransaction are actually duplicating
what genesisSpendable is doing, this is a simple replacement.
This step is required to allow splitting the genesis block in multiple modules (821).