Refactor a Ledger test to be more straightforward
It was actually quite hard to tell what this test intended to do due to the copy-pasting
and how much duplication there was.
By using TxBuilder and std.algorithm, we can drastically reduce the number of LoCs
while improving readability. The test description was also updated to reflect its intent.