Change default spend of extra blocks in unittests
The spending was too low to be able to use it
for many freeze transactions.

The QuorumPreimage test had to be changed because
of the way dice() works - even though the relative amount
of stake is the same between nodes, the random number
generator will in fact return different values.

For example:

```
dice(rnd, [10, 10, 10, 10, 10));
```

Returns different indices compared to:

```
dice(rnd, [100, 100, 100, 100, 100));
```

Even though the weight is the same.