Fix setting the is_nominating flag
This flag was erronously set to false on
the exit of the function, however the
nominating protocol is asynchronous.

This would lead to situations where:

- 8 transactions were received
=> Nomination started (async, nominate() returns early)
- 9th transaction received
=> Nomination started again!

This last nominating call would cause an increase
in the round number for the current slot index.

In practice this has not lead to any real-world
problems as multiple nominations are
actually quite normal.

However it could have caused issues with
the periodic catchup - as the node should not
add any blocks to its ledger while nominating.

in SCP it is the timer set up with setupTimer()
within NominationProtocol::nominate that should be
in charge of starting a second nominating round
if the first one has timed-out.

SCP gradually increases the timeout of every round,
starting from 1 second and then increasing it
for every additionally nominating round which failed.

Note that the transaction-based nomination trigger
will be replaced with a timer-based one.