Make the `registerListener` available for LocalRest
I registered a path for registering a listener manually in vibe.d.
But, in the LocalRest environment, we cannot use the REST path. So,
we should add an API endpoint for the Local environment. This commit
has derived new class named `TestNetworkClient` from the `NetworkClient`
so that it can set the API interface for the LocalRest environment.