Fix one more usage of toVec()
The nominate() routine is actually recursive and is
called again by a timer, and the timer has a delegate
which captures the two previous and next values.

These need to be on the C heap rather than in D's
GC heap because it might get collected by the GC
if it doesn't contain any roots to it.

Part of #723