Disable registerListener() code
The code was wrong because the request IP & port
is not the same as the listening IP & port of the
server which established the connection.

The net effect the previous code had was that
nodes would share their incoming connection
IP & ports to other nodes, resulting in an
explosion of addresses which each node tried to
(unsuccessfully) connect to.