Make the node call listenHTTP first
The runTask() call will spawn a new fiber and
and switch to the new fiber immediately.
However we want to start listening for incoming
requests before doing network discovery.

Using setTimer with a zero timeout (a.k.a. schedule())
makes the node open a listening port right away,
rather than waiting for the first discovery round
to complete.

It increases the node's connectivity to other nodes
which may want to establish a connection,
and consequently reduces CI failures.

Part of: #1117