Don't use a static variable for a status flag
In the unittests we have a mechanism of restarting a node
without shutting down the thread. Therefore any static
variables will remain with the last value which was set.

The `is_running` field needs to be reset when `AddressDiscoveryTask`
is first initialized, therefore it's been moved to become
a non-static field of the class.