Update Vibe.d & co to fix DNS timeout bug
One issue we are seeing in CircleCI is the lack of timeout for DNS queries.
Since the containers don't all start at the same time, some DNS queries get fired
before the containers are online, but an answer is never returned and the nodes
just hang. With this fix to Vibe.d, this should work correctly.