Tests: Fix timing issues with preimage test
The test expected the node to contain the commitment
preimage and then polled for the next preimage.

However if the preimage is revealed too quickly then
the commmitment will no longer be returned in the
getPreimage() call.

It is not necessary to check for the commitment preimage,
we can just check for the preimage itself (which the code
already does).