Simplify the implementation of quorum regeneration
Previously we used two distinct callbacks,
and then additionally we had to check for an edge-case
to prevent regenerating quorums twice.

However this can all be part of the same callback.

This removes the dependency inversion in the
Validator / FullNode constructors. And it also prevents
potential bugs when the Nominator is accessed from both
the onAcceptedBlock and onValidatorsChanged callbacks.

Now only onAcceptedBlock is used, which calls
regenerateQuorums internally.