Change the way pre-images are revealed
Before this commit, there was an timing issue where the pre-image
of a new validator cycle arrives before the new cycle does not
start yet which means a receiving node never get the pre-image again.
So, we need to change the way a validator reveals pre-images. We
have decided that validators keep revealing the pre-image set to be
revealed in the revealing period. After the period, validators start
and keep revealing new pre-image for a new period.