Schnorr: Implement extractNonce()
In the proposed signature scheme we want to
use the original (R, r) from the commitment,
and offset it with the preimage to derive
a new (R, r) pair.

For other nodes to be able to verify the
signature they must know the original R,
which they can extract from the enrollment
(commitment) signature.