Nominator: Implement schnorr multi-signatures
The scheme is:

- Node enrolls with rand seed (X) and uses the nonce (r) to sign,
  and includes r * G in the (R, s) signature.

At block height #1:
- r1 = r + x => for signing
- R1 = R + (x * G) => for validation

At height #2:
- r2 = r + preimage(x) => for signing
- R2 = R + preimage(x) * G => for validation

At height #3:
- r3 = r + preimage(preimage(x)) => for signing
- R3 = R + preimage(preimage(x)) * G => for validation

Implement a new signature scheme for envelopes

The envelope signature scheme can be derived from
the block signature scheme. We will not have to
store the R from the (R, s) signature pair
because we can simply derive it by offseting
it with the challenge.