Split KeyPair from config's Node to ValidatorConfig (#1197)
Currently, both full nodes and validators share a common `NodeConfig`, and are distinguished by the `is_validator` boolean.
However, a full node should not have a `KeyPair` (and other validator-related attributes) as its attribute.

As discussed, there are two ways of handling this issue:
1. Make a separate `ValidatorConfig` struct which contains attributes exclusive to a `Validator` **(current approach)**
2. Embed `NodeConfig` into a new `ValidatorConfig`

The current approach is ok, but gets messy in `Base.d` because `makeMainConfig` takes an extra `ValidatorConfig` as its parameter. The `ValidatorConfig` is also created for full nodes, just with initial values. The function could be overloaded to remove this parameter for full nodes.

In the 2nd approach, the main `Config` has to be templated to either take `NodeConfig` or `ValidatorConfig`. The disadvantage would be that more code changes will be required in agora as a whole.

Related to #816