Fix revelation of wrong pre-images after re-enrollment
There was a critical bug when `EnrollmentManager` reveals pre-images
in the case when a validator cycle does not end, but the validator has
requested new enrollment for the next cycle. In previous code, we `reset`
pre-images while creating new `Enrollment`, which means that the pre-images
of the current cycle has been replaced with new set of pre-images. So I make
the `populate` function only generate the initial random seed for enrollment.
After being a validator for the new cycle, the `populate` generates the whole
pre-images for the new cycle.