Replace (R, s) pair in envelope signature with (s)
SCP: Add a Schnorr signature to confirmed ballots

The signature signs the hash of the
proposed new block, in other words:

value_sig = sign(H(Block(ballot.value.deserialized)));

This allows us to interactively collect
signatures when the nodes vote to accept
& confirm a ballot.

The signature is a 32-byte scalar. We do not store
the R in (R, s) as it's already pre-commited.

We do not store the K as it's already set in the
SCPStatement itself.

BlockHeader: Change signature type to Scalar

We don't need to store the (R, s) pair,
as the R is known and derived from the
commitment and the revealed preimages.

Scalar: Add opSlice support

We want to be able to store the Scalar into
the `confirm_t`'s `value_sig` field,
which is an opaque array of 32 bytes.