Remove the `FakeClockBanManager`
Now that the BanManager has an instance of the Clock, we need not have
the FakeClockBanManager. So we remove the class and change the unittest
testing un-banning behavior to use the Clock of the BanManager. But we need
other class like `TestBanManager' due to the exception occured when `dumping`
the data into the file on LocalRest tests.