Add UTXO component in making pre-images
Our previous pre-image scheme relied only on the private key, a nonce, and
a fixed string. But the scheme made it unable for a node to enroll with
different frozen UTXOs because it made the same pre-images for the UTXOs.
So we have made it possible to enroll with different UTXOs by adding UXTO
hash to our pre-image scheme.