Implement a script execution Engine
The engine supports 4 different lock scripts:
    - Key => lock is a 64-byte public key,
             unlock is expected to be a signature.
    - KeyHash => lock is a hash of a 64-byte public key,
                 unlock is expected to be a pair of
                 <signature, public-key>.
                 This form may be used for better privacy.
    - Script => lock is a script that will be evaluated
                by the engine. The unlock script may either
                be empty or only contain stack push opcodes.
    - Redeem => lock is a <redeem-hash>, and unlock may
                only contain stack push opcodes, where the
                last push will be read as the redeem script.

In addition it supports arbitrary nesting of IF / ELSE
conditionals, and the set of opcodes as defined in
the `Opcodes.d` module.

Because the script lock contains a tag, it is possible
to extend and plug in a webassembly engine at a later point.

This commit does not integrate the Engine into Agora.
This should be part of a future commit.