Implement a Stack for scripting support
It uses a linked-list rather than a vector to avoid unnecessary copying
due to stomping prevention as the same item may be popped and later
pushed
to the stack. In addition, this makes it very cheap to copy the stack as
all internal items are immutable anyway.

The stack must be initialized with a set of size constraints,
the maximum size of the stack, and the maximum size of any one item
on the stack.