Schnorr: Demonstrate key cancellation attack
The current multi-sig schnorr signatures are only secure if
the owner of a public key proves that they own the associated
private key.

In our block multi-sig scheme this is a safe assumption,
because nodes sign an initial enrollment before becoming
a validator - thereby proving their ownership of the private key.

However, for users they would need to use an interactive
signature scheme where before agreeing on a common
X (public key aggregate) and R (public nonce aggregate),
they would need to prove ownership of their private key.

There are suggested alternatives which implement a secure
multi-sig scheme on top of Schnorr which do not require
the requirement for proving ownership of keys,
such as MuSig, which we should look into:

https://tlu.tarilabs.com/cryptography/musig-schnorr-sig-scheme/The_MuSig_Schnorr_Signature_Scheme.html

- https://tlu.tarilabs.com/cryptography/digital_signatures/introduction_schnorr_signatures.html#key-cancellation-attack
- https://blockstream.com/2018/01/23/en-musig-key-aggregation-schnorr-signatures/#:~:text=not%20secure.