Transaction: Add transaction-level time lock
The `unlock_height` field is used for a new consensus rule.
Transactions may only be included in blocks whose height
is >= `unlock_height`.

The new field is hashed and therefore changes all UTXOs
in the tests.

The consensus rules are moved into the next commit
to better split the diff.