Input: Add Input-level relative time-lock
The `unlock_age` field of an `Input` specifies the
minimum age difference between the referenced externalized
UTXO height, and the height at which the spending
transaction would be included in a block.

This changeset changes the hashing of the `Input`,
and therefore consensus rule changes were moved to
the next commit to make it easier to review.

Note that the Genesis block has not changed as it
has no inputs, therefore hashing is unaffected.