Implement SigHash-encoded signatures
SigHash was introduced to allow changing
how a signature from an Input is verified.

This is a strict requirement in order to implement:
- SIGHASH_NOINPUT for Eltoo (BIP118)
- SIGHASH_SINGLE for setting the fees for the final
  settlement transaction in the Eltoo protocol.

The basic payment lock types (LockType.Key / LockType.KeyHash)
expect a <signature, sighash> tuple in the
unlock byte array.

For scripts, the Engine now expects a tuple of <signature, sighash>
on the stack when verifying signatures.