Transaction: Add `seq_id` sesquence ID field
The sequence ID is used to be able to order
floating transactions as defined in the
Eltoo whitepaper. Transactions with an older
sequence ID cannot bind to transactions with
a newer sequence ID, therefore the latest
update / settlement transaction is easy to
determine.

Note that `seq_id` is not part of any
consensus rules, it is verified purely
by a script-level check in the execution
engine.