Implement consensus rules for time-locked transactions
Transactions which are not unlocked yet will not be
accepted by the node.

In the future this may be relaxed: A node may accept
a Transaction with a future unlock height into its
pool, but it should never accept it into a block.

Note that a transaction with a future unlock height
can be double-spent before it's externalized,
assuming the proper behavior of the transaction pool
(see #1332).

The `unlock_height` feature is primarily useful for
implementing HTLCs which will be necessary to support
multi-hop payment channels (Flash layer).