Implement consensus rules for Input time-locks
The `unlock_age` specifies the minimum age difference
between the height of the block of the spending
transaction, and the unlock height of the UTXO
it references.

This feature is needed for our Flash layer,
where transactions can attach to UTXOs only
after the utxo has been externalized.

Note that the `unlock_age` is compared to the
`Output's` unlock height, and not its externalize
height (unlock height is typically +1 of the
externalize height, for pay transactions.

This way an `unlock_age` of 0 and 1
have separate meanings:

`unlock_age` of 0 allows the `Input` to spend
the UTXO as soon as its unlock height is reached.

`unlock_age` of 1 allows the `Input` to spend
the UTXO when there was at least 1 block
externalized after the unlock height
of the UTXO being spent.

Fixed #243