Add support for SigHash.NoInput
This allows the implementation of
so-called floating transactions,
as described in the Lightning Network
and Eltoo whitepapers.

With this SigHash type, Outputs may
bind to any Inputs as long as signature
verification succeeds.

For this reason, it's important to
never reuse keys for lightning-layer
update / settlement key-pairs,
and instead use some form of key-derivation
as described in LN & Eltoo whitepapers.