Only re-enroll validators if there is none
The Ledger constructor was always re-enrolling the last few validators,
which means restarting a node would trigger an assertion failure.
This critical bug is worked around by only triggering this piece of code
when there is no validator (which should never happen).
Later on, this logic should be moved to catch-up.