Don't allow fiber switch before quorums are shuffled
A fiber context switch in the `onAcceptedBlock()` could
cause the nominating timer fiber to wake up again and
start nominating. But this could happen before the
new quorums were set, leading to an invalid quorum
state when nominating a new value.

Changed the call to `onAccepteBlock()` to be the last
call after the quorum generation has been completed.

Note: the proper fix is to implement a proper schedule()
mechanism in vibe.d & localrest.

Fixes #1408