ConsensusData nomination and validation
A `Validator` creates and nominates a `ConsensusData` struct with the
indices of the UTXOs of validators that are missing their preimages
(a.k.a. MPV). Whea a `Validator` receives a `ConsensusData` struct it
will try to verify that the `MPV` in the `ConsensusData` is the same
as the local MPV.