Add support for non-interactive threshold signatures
The `CHECK_MULTI_SIG` and `VERIFY_MULTI_SIG` enable
writing script locks which require a threshold
number of valid signatures for the provided public keys.

In the following text:
  N = number of sigs
  M = number of keys

It supports any combination of `N / M` where `N <= M`
and where the number of keys or signatures does not
exceed 5.

The limit is in place due to the relatively expensive
signature validation process, as each signature is
compared with every key until it either validates or
until the next key is about to be tested.