Use exceptions instead of assertions
The `fromString()` routines do not know whether
the input string was created by the programmer
or whether it was read from the network.

In particular vibe.d uses the fromString() routines.

Therefore the function cannot assume that a badly-formed
string is a coding error. The call site can
decide to turn an exception into an assertion
if it needs to.

Fixes #1399