Work around buggy libexecinfo by defining our own traceHandler
On Alpine Linux, which uses Musl libc, there is no backtrace() and co.
As a result, to provide trace information when an exception is thrown,
druntime was made to rely on libexecinfo, which provides a backtrace-like interface.

However, it turned out that this library is buggy and routinely leads to SEGV
in various pieces of code. Hence, we are in the process of adding support
for libunwind to druntime, which is the industry standard when it comes to frame inspection.

The work upstream is almost complete, and the results we've seen are very
encouraging. However, the support will only land in v2.096.0, which means
LDC v1.26.0, which is two release away, estimated to be around 4 months from now.
It is likely that this support can be back-ported to LDC v1.25.0 with a little effort,
which would put the ETA at 1 month, much more reasonable.

Given this uncertain schedule, we want to work around the issue by copying over the support
from the druntime PR. The workaround is rather crude: we simply define a trace handler
that prints addresses and function names. This means we loose file/line information temporarily,
but Agora should be much more stable.

If the missing file/line information turns out to be problematic,
we could get it back by copying over the relevant parts of druntime,
as they have been made to be self-contained by our previous refactorings.

As a last note, this also means we force users to link in `libunwind` on other POSIX platforms.
It's not an issue on OSX, as it's distributed with the development environment,
but will create another constraint on Linux.