Introduce getPreimages() API endpoint
We need to get all the preimages of validators in a specified range
of `enrolled heights` in the catch-up process. It's  because checking
validity of a block needs preimages of validators in order to
generate random seed. So we introduce the getPreimages API endpoint
which returns the preimages from the start `enrolled height` to the end
'enrolled height`.