Add TrackingGC, a runtime, opt-in way to track memory allocations
This integrate with the "Tracy" library to provide us with more
insight into what is happening inside our program.
As we currently have a memory leak, this should help us detect
the places where allocations take places (and are never released).
Note that this copies the upstream GC code, and add instrumentation
when needed. This is necessary as otherwise GC frees done after
collection would not be recorded (the GC doesn't call `GC.free`
on dead objects, it would be too expensive).