SpyValidator: Remove unused curr_time and re-order to match superclass
The argument 'curr_time' was not used anywhere, most likely because
it is obsolete since the introduction of the `Clock` type.
In some places, the `test_start_time` argument also has been moved
directly after `txs_to_nominate`, to match the superclass,
and remaining arguments are put at the end.