Make `IBlockStorage.readLastBlock` return a `Block`
Like `readBlock`, the cost of an `Exception` for `readLastBlock`
is minimal compared to the IO to do, and it doesn't make
sense to throw in the lower layer and return a boolean in the
upper layers, so just accept that it won't be nothrow.

Not to mention that two out of three usages would lead to an assert failure,
so the code currently doesn't handle this properly anyway.