Flash: Implement Channel class
This class is responsible for maintaining a
single 2-party channel. There may be multiple
channels open between two parties, but the Channel
class only cares about a single channel instance.

The channel's ID is derived from the funding transaction's
UTXO hash. The channel becomes open once the funding
transaction is externalized in the blockchain. The channel
becomes closed once the closing transaction is externalized,
or may alternatively become closed after the externalization
of the trigger / update and settlement transactions.