Flash: Implement Invoice definition
The invoice is used to route HTLC payments
across the network. The node shares the hash
of a preimage to the party which intends to
pay for the invoice.

The destination node can redeem the payment
if the reveal the secret to the counter-party
of the channel through which the payment
was routed.

The secrets are then revealed in the reverse
path of the payment, thereby updating all
the balances and collecting fees.