Flash: Implement encrypted onion packets
These packets ensure that each hop node in
the payment route only knows to decrypt
their own information and may only
route the payment to the next channel.

Note: Currently the size of the onion is not static.
This should be improved for privacy / censorship
reasons (See LN RFC spec).