BanManager: Only ban after 1k failed requests
Currently the BanManager is a tad too eager to ban nodes,
which is causing many issues, both in the live system and
the test-suite. While the underlying reason it's getting banned
might persist and still trigger the ban, it is much likely
to be less spurious with a higher limit.