Ledger: Move IO/side effects to `acceptBlock`, make `addValidatedBlock` Ledger-only
There was a split of responsibilities between what `acceptBlock` and
`addValidatedBlock` were doing. `acceptBlock` was mostly forwarding
to `addValidatedBlock`, but also logging stats, except for one metric.

Because the node needed to potentially catch-up on startup,
a boolean (externalized) had to be added for this to work.
Since the boolean was only ever set to false a single time,
it's better to remove it, and instead have the validation duplicated
in the startup code.

In order for this to work, we also needed to carve the side-effect
performing code out of `addValidatedBlock`, as trying to save
an already saved block would have triggered an assertion.

The result is a public interface which does all the IO operations,
and forward to a private interface which does what the ledger
is supposed to do: update its state.