Use crypto lib including hash of array lengths
To prevent vulnerabilities when hashing adjacent arrays in structs
we include a hash of the length of those variable length arrays.
In cases where we want to hash an array of ubyte which is not fixed
length but we know it should be hashed without the length we can wrap the
ubytes in a HashNoLength struct.