NetworkManager: Remove buggy getAddress/isOurOwnAddress
isOurOwnAddress was not working as intended, since it relied on getAddress,
which itself relied on the config value of 'node.address', which is not
our own address, but rather a netmask of interfaces to listen to.
The most value for this field is actually '0.0.0.0',
meaning we're listening on all available interfaces,
so we would never attempt to connect to 0.0.0.0.

This also uncovered a bug in LocalRest that triggers the submodule update.