Add `--initialize` and an endpoint to validate the config file
This is the skeleton of the setup interface.
When `--initialize` is passed to Agora, it enters a "setup"
step where it listens on 127.0.0.1:2827 for requests.

It will serve a single-page-application (SPA) codenamed Talos,
which at the end of the setup process will call the `/create` endpoint
with the generated config file. If the config file passes validation,
it will be written to disk, the setup interface / Talos be shut down,
and Agora will start.