Block validation: Replace `is[Not]Valid` with `assertValid`
isValid was originally intended as a mean to transition from the bool-returning isValid
to the (at the time, newly introduced) string-returning isInvalidReason.
Since this has long been done, isValid is only used in assert nowadays,
and has been modified to provide better UX for developers when a test fail.
Changing it to 'assertValid' makes sure that this new meaning is well-reflected.