Tasks: Separate Vibe.d implementation from the interface
Ideally, we should be able to completely ignore Vibe.d when
building the test target, provided all usages of Vibe.d have
been properly abstracted.
This is currently not possible for a number of reasons,
including the fact that the API and the implementations depend on Vibe.d.
Separating the `TaskManager` gives us one less reason.