CI: Add retry hack to version_build & fix existing
We get an odd 'no such file or directly' error on the CI sometimes,
which seems to come from a race condition in dub.
The retry hack used for the C++ build script was broken because it
used `sleep || ...`, which should have been `sleep && ...` (sleep doesn't fail).
So fix it, and apply the same hack to `version_dub`.