UnitTest runner: Stop early in single threaded mode
Currently, we have some flakey test in the CI, so we have a retry strategy in place.
Namely, we try to run the unittest binary twice.
However, our retry strategy is defeated by our module runner,
which will run all the tests regardless. This isn't very helpful,
especially considering the amount of output a single test failure produces.
So if a single threaded mode fails, consider it important enough to warrant
forcing the developer forcing it before continuing.