Remove unnecessary `convert_base` parameter
The parameter could be removed with some refactoring. One value for testing
was removed. The removed value for testing is actually not needed because
it's just for checksum testing and we have to check the validity of encoded
or decoded value. So we removed the value and made the encoding/decoding
functions simpler.