Bech32: Use Exceptions instead of Logger
Since the implementation is still being tested,
we want to have as much information as possible if an error happens,
hence why rich error through the Logger was useful.
However, it adds some overhead and makes the implementation depends on Ocean.
As we want to upstream it, we move it to Exception as a first step.