Set default values for retry and banning
The configuration of retry and banning is more related than expected.
We should wait some time to ban any unanswering node after many retries
and failed requests. We set the waiting time to about 3 hours. Actually,
the total time of multiplication of `retry_delay`, `max_retries`,
`timeout`, and `max_failed_requests` could be one and a half hours. We
mostly refered to the values of `infrastructure` because it is used
for the live test system. Anyway, one hour could be the best time to
wait before banning a bad node.