Reorder the database into two main databases
Currently we have as many databases as we have classes,
most of them with a single table in them.
This refactoring shifts Agora to only two databases,
with the state database being the most important one,
and the cache database being expandable.
This should make inspecting the state of a node on a live system
much easier, as one won't need to switch between prompts and files.