Move signature verification from validation to Leder
Signature verification requires way too much context to be in the validation package.
The validation package was originally intended as a place to store syntactic checks
(checks that require little to no context), but it has grown into a place where all
the checks are performed, which led to a maze of dependencies.
Moving things back to the Ledger clears the coast a little.