Block signature validation: Clean up & improve logs
This tame the amount of logging going on in the signature validation.
It should now only log once if there's not enough validators,
twice if there's an error (the number of validator and the error),
or just once without an error (the number of signers).
Also turn some logging into assertions, as they should never happen,
and if they do it's a programming error.