Use preimage to generate the block signature noise
This is to prevent nodes from signing more than one block at a certain
height. The preimage, which is unique per block height, is used to
generate the signature noise. If the same signature noise is used
multiple times then the node's private key can be calculated. There is a
unit test to demonstrate this calculation.