Unit test ValidatorCount improved
Firstly this test was missing the assertion that only 1 block was
returned after attempting to fetch 2 at the start of the test.
Secondly it only needs to check that the height is correct to confirm
that a block was not added when there are no validators available for
the next block.