Fix a block retrieval bug
Scenario:

- NetworkClient.getBlocksFrom() retrieves
  a set of blocks [10, 11, 12, 13, 14],
  but the last block #14 was invalid.
- The Ledger was updated to point to block #13.
- This code was problematic:
  if (!onReceivedBlocks(blocks, preimages))
      continue LNextNode;
  block_height += blocks.length;

Then `block_height` would not get updated,
so when we tried reading from the next node.
But the blocks would be rejected by the ledger
because we would try to start with height
`#10` again.
`FullNode.addBlocks()` would fail here.

In the end this shouldn't really be a big
problem because the block retrieval is done
in a periodic timer in `FullNode.startPeriodicCatchup()`,
and in it we use the Ledger's latest block height.

This is merely an optimization fix.