Check if async nominating already
This is related to SCP nomination where we are already in process of
nominating but the checkNominate is called by the timer. The check if we
are already in the process of a nomination was removed when a check that
we only confirm once was added. This adds back the check.