Fix addBlocks() to not reject valid blocks
There was a bug with the way `addBlocks` short-circuited
and returned early.

If any of the preimages were rejected (for *any* reason),
further blocks would be rejected.

However, there are at least two cases where the internal
preimage state could mess with `addBlocks` and cause
it to return early:

A) We already have a duplicate preimage internally,
   from one of the previous receivePreimage() calls.
B) The preimage is for a newer enrollment which is in one of the
   `blocks` which we haven't read from yet. We read the
   blocks sequentally but the preimages were being
   added for every block height. As soon as one block
   was accepted, the second one would be rejected because
   of the short-circuit behavior (duplicate preimage found).