Use expected time offset for block nominations
To prevent many different block offsets in nominations which will often
prevent convergence, we use the expected offset. This is first step to
make the block creation more stable.
When deciding if to nominate we should take into account the block
time offset tolerance as the network clocks are not currently synched.