Tests: Fix wrong tx check assumptions
The code was sending tx's to nodes
and then immediately after queried
if the nodes accepted those tx's
by using the hasTransactionHash() API.

However there is a race condition here.
The nodes will begin a new nomination
round and may quickly externalize a new
block, which will remove those txs
from the tx pool.

Therefore the tests would sometimes
fail if the nodes externalized the
blocks faster than the main thread
called hasTransactionHash().

As an alternative a test-only
hasAcceptedTxHash() API has been
added which checks if the given
tx hash was *ever* accepted into
the pool, even if it was later
removed by externalization of
a new block.

hasTransactionHash() should only
be used when we know for sure
a new block should not be
externalized.