Flash: Implement onChannelNotify()
This API is used by wallets to get
messages about state changes of
a channel.

The openNewChannel() API is no longer
a blocking API, but instead returns
errors in case there are duplicate
funding UTXOs for existing pending
or open channels.

To get notifications when a channel
is open (or rejected by counter-party),
the onChannelNotify() Listener API
should be used.

The Node uses an internal queue
of pending channels issued by
openNewChannel(). It uses a periodic
fiber to initiate the setup
stage for the channel.