Flash: Move peer initialization out of ctor
Previously we used a workaround where we initialized
the UpdateSigner late when the channel was loaded
from storage (because the event loop wasn't ready).

But this also caused issues where we queried
the update manager before it was initialized.

To simplify code, just pass the peer in the
call to collectSignatures() and simply
instantiate the UpdateManager in the ctors.