Use expected time offset for block nominations
To prevent many different block offsets in nominations which will often
prevent convergence, we use the expected offset. Blocks are allowed to
be made with only offset at least the value of the previous block.
Note that the concept of a time tolerance is no longer included. As long
as enough nodes agree on the time offset then it will be used.