Use height for pre images not distance
This changes the database field `distance` to `height` in the
`validator` table as it now is the block height for the `pre-image`.
This makes it much easier to work with preimages.
The `active` flag is also removed as we can deduce if an enrollment is
active from the `enrollment_height`. We keep all the previous enrollment
cycle records so that we know which validators were active during
previous cycles. This is needed for verifying signatures that we recieve
after a new cycle has started.