Change FullNode.addBlocks to not trigger a spurious error
This method would sometimes show an error because we were trying to add
a block at the same height as the Ledger to it.
Sometimes, when things were very delayed, we even tried to add older blocks.
The exact message was 'Block: Height is not one more than previous block'
from the block validation routines. This can happen if we get the response
to our 'getBlockFrom' after a new block has been externalized.
Additionally, a redundant local variable (last_height) has been removed.