Part of 1972: Do not repeat node logs in assert handler
As explained in issue 1972, we were sometimes printing node logs twice.
Since tests nowadays run in single-threaded mode, we can take advantage
of the knowledge we're in the main thread.
In the future, when we move back to run unittests in parallel,
we might need to set a flag from the unittest runner.