Serve Talos web interface from Agora
The regular Agora doesn't start immediately after the setup is complete,
instead the user have to start the node themselves.
This was changed to prevent misuse or misconfiguration.

Additionally, some aspects of the interface are still a bit rough or unfinished,
notably when it comes to error messages and what is configurable.
This will be improved in the near future.