Check test GenesisBlock keys
When we use the genesis enrolled node keys in unit tests we want the
index of the node key to match the index of the matching enrollment.
In some tests we need to know which indexes to wait for blocks or
preimages. Having the genesis_validator_keys indexes matching the
enrollments in the test GenesisBlock helps.