Talos: Upgrade node-gyp to v7 (and more)
We were using node-gyp v3.8.0, which has a bug preventing it to be used with Python3
However, we build on Alpine Linux, which requires Python3.

As the bug was showing up while building node-sass,
the later was also updated to a compatible version.
It also helps as the node-sass version was ancient and did not
work with all recent node.js versions.

However, as `react-scripts` depended on `sass-loader` v8,
and `sass-loader` depends on `node-sass` v4.x.x, we also had to bump `react-scripts`.
The first version to support `sass-loader` v11 was v4.0.2 and v4.0.3 is the latest
version as of the time of this commit.

This then led to https://github.com/facebook/create-react-app/issues/9937
which is worked around by simply moving the assets.