CI: Install libsodium directly on OSX
When the CI was put in place, running brew update was extremely slow (3+ minutes).
At the time, it was dominating the CI time. The solution we used was to directly
download the bottle from bintray and install / link it, and it worked well.
However, Bintray has shut down on May 1st and this script is now broken.
As Homebrew has migrated to use Github packages directly, and there has been a few
releases of homebrew since then, brew upgrade is long longer necessary,
and the download is likely to be much faster.