Unit test refactor to remove `multiSigTestBlock`
The `multiSigTestBlock` unit test function was used to sign a Block
in unit tests with code similar to the prodution code. For some tests it
is required to mimic the behavior of multiple nodes signing a block.
However where posible the real code should be used and only a thin layer
of test code which can access the `WK.keys` (well known keys for unit
tests) and call the real implementation.