Unit test refactor to remove `getTestRandomSeed`
The `getTestRandomSeed` function was hashing `Hash.init` to use in unit
tests to avoid the assertions in `block.isInvalidReason`. This leads to
test code which is harder to reason about. As a first step and to make
it clearer what is being tested this removes the `getTestRandomSeed`.
The assertion in the `block.isInvalidReason` is not included if running
`version (unittest)` mode. This is still not ideal but I think is a step
in the right direction.